lab-multimodal/requirements.txt
dotenv
azure-ai-inference